Config layout refactor: centralize config file for documentation
