+++ common/config/sasl.go
@@ -18,7 +18,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package auth
+package config
 
 type (
 	// SASL describe SASL configuration (for Kafka)
