+++ common/config/tls.go
@@ -18,7 +18,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package auth
+package config
 
 type (
 	// TLS describe TLS configuration (for Kafka, Cassandra, SQL)
